bodyprog: rodata migrate `GameState_MainMenu_Update`
Could only make this work with new file split, odd
