+++ src/bodyprog/bodyprog_8003AB28.c
@@ -1,0 +1,2522 @@
+#include "game.h"
+
+#include <libetc.h>
+#include <libpad.h>
+#include <ctype.h>
+#include <strings.h>
+
+#include "bodyprog/bodyprog.h"
+#include "bodyprog/credits.h"
+#include "bodyprog/demo.h"
+#include "bodyprog/item_screens.h"
+#include "bodyprog/math.h"
+#include "bodyprog/memcard.h"
+#include "bodyprog/player_logic.h"
+#include "main/fsqueue.h"
+#include "main/mem.h"
+#include "main/rng.h"
+#include "screens/stream/stream.h"
+
+const s32 g_rodataPad_8002547C = 0;
+
+void GameState_MainMenu_Update() // 0x8003AB28
+{
+    s32 nextGameStates[5] = // 0x80025480
+    {
+        GameState_SaveScreen, 
+        GameState_DeathLoadScreen,
+        GameState_MovieOpening,
+        GameState_OptionScreen,
+        GameState_MovieIntro
+    };
+
+    s32           playIntroFMV;
+    s32           var_v0;
+    s32           var_2;
+    s_ShSavegame* save0;
+    e_GameState   prevState;
+
+    func_80033548();
+	/** Checks if after being idle for a while in the title screen it will reproduce
+	 * a demo gameplay or the intro FMV, if the next value from `g_Demo_ReproducedCount`
+	 * is a value that jumps in 3 (starting to count from 0) then it will reproduce the
+	 * intro FMV if not it will reproduce a demo gameplay
+	 */
+    playIntroFMV = ((g_Demo_ReproducedCount + 1) % 3) != 0;
+
+    if (g_GameWork.gameStateStep_598[0] == 0)
+    {
+        g_MainMenuState = 0;
+        
+        if (playIntroFMV != 0)
+        {
+            g_SysWork.flags_2298 = 0x20; // This disables the player control
+        }
+        else
+        {
+            g_GameWork.gameStateStep_598[0] = 1;
+        }
+    }
+
+    switch (g_MainMenuState)
+    {
+        case MenuState_Start:
+            g_GameWork.background2dColor_R_58C = 0;
+            g_GameWork.background2dColor_G_58D = 0;
+            g_GameWork.background2dColor_B_58E = 0;
+
+            Gfx_ClearRectInterlaced(0, 32, SCREEN_WIDTH, FRAMEBUFFER_HEIGHT_INTERLACED, 0, 0, 0);
+            Gfx_Init(SCREEN_WIDTH, 1);
+
+            g_IntervalVBlanks = 1;
+            g_Gfx_ScreenFade  = 6;
+            D_800B5C30        = 0x2000;
+            g_MainMenuState++;
+
+        case MenuState_Main:
+            if (playIntroFMV != 0)
+            {
+                Demo_StartUp();
+
+                if (g_GameWork.gameStateStep_598[0] == 1 && g_SysWork.timer_20 == 0)
+                {
+                    g_Demo_ReproducedCount++;
+                }
+
+                if (g_GameWork.gameState_594 == GameState_MainLoadScreen)
+                {
+                    g_Demo_ReproducedCount++;
+                }
+            }
+
+            D_800A9A7C = 12;
+
+            if (g_GameWork.savegame_90.playerHealth_240 > 0)
+            {
+                D_800A9A7C = 14;
+            }
+
+            if (D_800BCD28 > 0)
+            {
+                D_800A9A7C |= (1 << 0) | (1 << 1);
+                
+                if (D_800A9A88 < D_800BCD28 && g_MainMenu_SelectedOptionIdx != 0)
+                {
+                    g_MainMenu_SelectedOptionIdx = 1;
+                }
+            }
+            else if (D_800A9A88 > 0)
+            {
+                while(!(D_800A9A7C & (1 << g_MainMenu_SelectedOptionIdx)))
+                {
+                    g_MainMenu_SelectedOptionIdx++;
+                }
+            }
+
+            D_800A9A7C |= D_800A9A7C << 5;
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & (ControllerFlag_LStickUp | ControllerFlag_LStickDown))
+            {
+                Sd_EngineCmd(0x519);
+                g_GameWork.gameState_594 = GameState_MainMenu;
+
+                if (g_GameWork.gameStateStep_598[0] != 1)
+                {
+                    g_GameWork.gameStateStep_598[0] = 1;
+                    Fs_QueueReset();
+                }
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & ControllerFlag_LStickUp)
+            {
+                g_MainMenu_SelectedOptionIdx += 5;
+                while(!(D_800A9A7C & (1 << --g_MainMenu_SelectedOptionIdx)))
+                {
+                }
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & ControllerFlag_LStickDown)
+            {                
+                while(!(D_800A9A7C & (1 << ++g_MainMenu_SelectedOptionIdx)))
+                {
+                }
+            }
+
+            g_MainMenu_SelectedOptionIdx %= 5; // This is done for avoid overflow
+
+            if (g_ControllerPtrConst->btnsClicked_10 & g_GameWorkPtr->config_0.controllerConfig_0.enter_0)
+            {
+                g_GameWork.gameState_594 = GameState_MainMenu;
+
+                if (g_GameWork.gameStateStep_598[0] != 1)
+                {
+                    g_GameWork.gameStateStep_598[0] = 1;
+                    Fs_QueueReset();
+                }
+
+                g_Gfx_ScreenFade = 2;
+                g_MainMenuState++;
+
+                if (g_MainMenu_SelectedOptionIdx < 2u)
+                {
+                    Sd_EngineCmd(0x501);
+                }
+                else
+                {
+                    Sd_EngineCmd(0x51B);
+                }
+
+                switch (g_MainMenu_SelectedOptionIdx)
+                {
+                    case 1: // Quick load
+                        if (g_GameWork.savegame_90.playerHealth_240 > 0)
+                        {
+                            g_GameWork.savegame_30C = g_GameWork.savegame_90;
+                        }
+                        else
+                        {
+                            GameFs_SaveLoadBinLoad();
+                        }
+
+                        func_80035178();
+                        g_SysWork.flags_2298 = 0x10;
+                        GameFs_MapLoad(g_SavegamePtr->mapOverlayId_A4);
+                        break;
+
+                    case 0: // Load save and load menu
+                        GameFs_SaveLoadBinLoad();
+                        break;
+
+                    case 2:
+                        g_Gfx_ScreenFade = 0;
+                        g_MainMenuState  = MenuState_DifficultySelector;
+                        break;
+
+                    case 3: // Load options menu
+                        GameFs_OptionBinLoad();
+                        break;
+
+                    case 4:
+                        break;
+                }
+            }
+
+            D_800A9A88 = D_800BCD28;
+
+        default:
+            break;
+
+        case MenuState_DifficultySelector:
+            if (playIntroFMV != 0)
+            {
+                Demo_StartUp();
+
+                if (g_GameWork.gameStateStep_598[0] == 1 && g_SysWork.timer_20 == 0)
+                {
+                    g_Demo_ReproducedCount++;
+                }
+
+                if (g_GameWork.gameState_594 == GameState_MainLoadScreen)
+                {
+                    g_Demo_ReproducedCount++;
+                }
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & (ControllerFlag_LStickUp | ControllerFlag_LStickDown) ||
+                g_ControllerPtrConst->btnsClicked_10 & (g_GameWorkPtr->config_0.controllerConfig_0.enter_0 |
+                                                        g_GameWorkPtr->config_0.controllerConfig_0.cancel_2))
+            {
+                g_GameWork.gameState_594 = GameState_MainMenu;
+                
+                if (g_GameWork.gameStateStep_598[0] != 1)
+                {
+                    g_GameWork.gameStateStep_598[0] = 1;
+                    Fs_QueueReset();
+                }
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & ControllerFlag_LStickUp)
+            {
+                var_v0 = 2;
+
+                if (g_MainMenu_NewGameSelectedDifficultyIdx > 0)
+                {
+                    var_v0 = g_MainMenu_NewGameSelectedDifficultyIdx - 1;
+                }
+
+                g_MainMenu_NewGameSelectedDifficultyIdx = var_v0;
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & ControllerFlag_LStickDown)
+            {
+                var_2 = 0;
+
+                if (g_MainMenu_NewGameSelectedDifficultyIdx < 2)
+                {
+                    var_2 = g_MainMenu_NewGameSelectedDifficultyIdx + 1;
+                }
+
+                g_MainMenu_NewGameSelectedDifficultyIdx = var_2;
+            }
+
+            if (g_ControllerPtrConst->btnsPulsed_18 & (ControllerFlag_LStickUp | ControllerFlag_LStickDown))
+            {
+                Sd_EngineCmd(0x519u);
+            }
+
+            if (g_ControllerPtrConst->btnsClicked_10 & g_GameWorkPtr->config_0.controllerConfig_0.enter_0)
+            {
+                Game_SavegameInitialize(0, g_MainMenu_NewGameSelectedDifficultyIdx - 1);
+                func_80035178();
+
+                g_SysWork.flags_2298 = 4;
+
+                GameFs_MapLoad(0);
+                GameFs_StreamBinLoad();
+                Sd_EngineCmd(0x501);
+
+                g_Gfx_ScreenFade = 2;
+                g_MainMenuState  = 4;
+            }
+            else if (g_ControllerPtrConst->btnsClicked_10 & g_GameWorkPtr->config_0.controllerConfig_0.cancel_2)
+            {
+                Sd_EngineCmd(0x51A);
+                g_MainMenuState = 1;
+            }
+            break;
+
+        case MenuState_LoadGame:
+        case MenuState_NewGameStart:
+            if ((g_Gfx_ScreenFade & 0x7) == 5)
+            {
+                Gfx_ScreenRefresh(320, 0); // Old idb `Sys_GFXReinit_800323C8(width, interlace_flag)`.
+                Fs_QueueWaitForEmpty();
+
+                if (g_GameWork.savegame_90.playerHealth_240 > 0)
+                {
+                    nextGameStates[1] = 10;
+                }
+
+                if (g_MainMenu_SelectedOptionIdx == 2)
+                {
+                    Chara_PositionUpdateFromParams(g_MapOverlayHeader.mapAreaLoadParams_1C);
+                }
+
+                func_8002E830();
+
+                prevState                       = g_GameWork.gameState_594;
+                g_GameWork.gameStateStep_598[0] = prevState;
+                g_GameWork.gameState_594        = nextGameStates[g_MainMenu_SelectedOptionIdx];
+                g_SysWork.timer_1C              = 0;
+                g_GameWork.gameStatePrev_590    = prevState;
+                g_GameWork.gameStateStep_598[0] = 0;
+                g_SysWork.timer_20              = 0;
+                g_GameWork.gameStateStep_598[1] = 0;
+                g_GameWork.gameStateStep_598[2] = 0;
+                SysWork_StateSetNext(SysState_Gameplay);
+            }
+            break;
+    }
+
+    if (g_ControllerPtrConst->btnsHeld_C != 0)
+    {
+        g_SysWork.timer_20 = 0;
+    }
+
+    if (playIntroFMV == 0)
+    {
+        switch (g_GameWork.gameStateStep_598[0])
+        {
+            case 1:
+                if (g_SysWork.timer_20 > 1740)
+                {
+                    GameFs_StreamBinLoad();
+                    g_GameWork.gameStateStep_598[0]++;
+                }
+                break;
+
+            case 2:
+                if (Fs_QueueGetLength() == 0)
+                {
+                    g_Demo_ReproducedCount++;
+
+                    g_GameWork.background2dColor_R_58C = 0;
+                    g_GameWork.background2dColor_G_58D = 0;
+                    g_GameWork.background2dColor_B_58E = 0;
+
+                    Game_StateSetNext(GameState_MovieIntro);
+                }
+                break;
+        }
+    }
+
+    if (g_GameWork.gameState_594 == GameState_MainMenu)
+    {
+        Gfx_MainMenu_BgDraw();
+        func_8003B560();
+
+        if (g_MainMenuState < 3)
+        {
+            Gfx_MainMenu_MainTextDraw();
+            return;
+        }
+
+        Gfx_MainMenu_DifficultyTextDraw(g_MainMenu_NewGameSelectedDifficultyIdx);
+        return;
+    }
+
+    *(s32*)0x1F800000 = 0x200000;
+    *(s32*)0x1F800004 = 0x01C00140;
+    ClearImage2((RECT* )0x1F800000, 0u, 0u, 0u);
+    Gfx_Init(0x140u, 0);
+}
+
+void func_8003B550() // 0x8003B550
+{
+    g_MainMenu_SelectedOptionIdx = 1;
+}
+
+void func_8003B560() {}
+
+void Gfx_MainMenu_MainTextDraw() // 0x8003B568
+{
+    #define STR_POS_X_BASE 158
+    #define STR_POS_Y_BASE 184
+
+    s32 i;
+
+    for (i = 0; i < 5; i++)
+    {
+        if (D_800A9A7C & (1 << i))
+        {
+            Gfx_StringSetPosition(STR_POS_X_BASE - D_800254EC[i], STR_POS_Y_BASE + (i * 20));
+            Gfx_StringSetColor(ColorId_White);
+
+            if (i == g_MainMenu_SelectedOptionIdx)
+            {
+                Gfx_StringDraw(&D_800254F4, 99);
+            }
+            else
+            {
+                Gfx_StringDraw(&D_800254F8, 99);
+            }
+
+            Gfx_StringDraw(D_800A9A8C[i], 99);
+
+            if (i == g_MainMenu_SelectedOptionIdx)
+            {
+                Gfx_StringDraw(&D_800254FC, 99);
+            }
+
+            Gfx_StringDraw(&D_80025500, 99);
+        }
+    }
+}
+
+void Gfx_MainMenu_DifficultyTextDraw(s32 arg0) // 0x8003B678
+{
+    s32 i;
+
+    for (i = 0; i < 3; i++)
+    {
+        Gfx_StringSetPosition(158 - D_8002551C[i], 204 + (20 * i));
+        Gfx_StringSetColor(ColorId_White);
+
+        if (i == arg0)
+        {
+            Gfx_StringDraw(&D_800254F4, 99);
+        }
+        else
+        {
+            Gfx_StringDraw(&D_800254F8, 99);
+        }
+
+        Gfx_StringDraw(D_800A9AA0[i], 99);
+
+        if (i == arg0)
+        {
+            Gfx_StringDraw(&D_800254FC, 99);
+        }
+
+        Gfx_StringDraw(&D_80025500, 99);
+    }
+}
+
+void Gfx_MainMenu_BgDraw() // 0x8003B758
+{
+    if (g_SysWork.sysState_8 == 0)
+    {
+        g_SysWork.sysState_8     = 1;
+        g_SysWork.timer_24       = 0;
+        g_SysWork.sysStateStep_C = 0;
+        g_SysWork.field_28       = 0;
+        g_SysWork.field_10       = 0;
+        g_SysWork.timer_2C       = 0;
+        g_SysWork.field_14       = 0;
+        func_8003BCF4();
+    }
+    Gfx_BackgroundSpriteDraw(&D_800A9014);
+    func_8003BC8C();
+}
+
+void func_8003B7BC() // 0x8003B7BC
+{
+    // Can't be `s32*` since 462 doesn't divide by 4, so I'm guessing it's `s8`.
+    s8* s0 = 0x801E2432;
+
+    memset(s0, 0, 462);
+    D_800BCDE0 = s0;
+}
+
+u32 func_8003B7FC(s32 idx) // 0x8003B7FC
+{
+    u8  idx0 = D_800BCDE0[idx];
+    u32 val = D_800A9AAC[idx0];
+
+    if (idx < 210)
+    {
+        return 0x3A000000;
+    }
+
+    return val;
+}
+
+PACKET* func_8003B838(GsOT* ot, PACKET* packet) // 0x8003B838
+{
+    s32      yOffset;
+    s32      i;
+    s32      j;
+    s32      color0;
+    s32      color2;
+    s32      color3;
+    s32      color1;
+    POLY_G4* poly;
+
+    for (i = 10; i < 21; i++)
+    {
+        color1 = func_8003B7FC(21 * (i - 1));
+        color3 = func_8003B7FC(21 * i);
+
+        for (j = 1; j < 21; j++)
+        {
+            color2 = color3;
+            color0 = color1;
+
+            color1 = func_8003B7FC(j + (21 * (i - 1)));
+            color3 = func_8003B7FC(j + (21 * i));
+
+            poly = packet;
+            setPolyG4(poly);
+
+            yOffset = (i - 1) * 24;
+
+            setXY4(poly,
+                   (-176 + 16 * j), yOffset - 208,
+                   (-160 + 16 * j), yOffset - 208,
+                   (-176 + 16 * j), yOffset - 184,
+                   (-160 + 16 * j), yOffset - 184);
+
+            *((u32*)&poly->r0) = color0;
+            *((u32*)&poly->r1) = color1;
+            *((u32*)&poly->r2) = color2;
+            *((u32*)&poly->r3) = color3;
+
+            addPrim(ot, poly);
+            packet += sizeof(POLY_G4);
+        }
+    }
+    return packet;
+}
+
+void func_8003BA08() // 0x8003BA08
+{
+    PACKET*   packet;
+    GsOT_TAG* tag;
+
+    tag    = g_ObjectTable1[g_ObjectTableIdx].org;
+    packet = func_8003B838(&tag[6], GsOUT_PACKET_P);
+    SetDrawMode((DR_MODE*)packet, 0, 1, 0x2A, NULL);
+    addPrim(&tag[6], packet);
+    GsOUT_PACKET_P = packet + sizeof(DR_MODE);
+}
+
+// Matched in decomp.me, but inserted has missmatch.
+#ifdef NON_MATCHING
+void func_8003BAC4() // 0x8003BAC4
+{
+    s32 idx;
+    s32 i;
+    s32 value;
+    s8* ptr;
+    u8* ptr1;
+    s8* ptr2;
+
+    ptr = D_800BCDE0;
+    ptr1 = ptr + 441;
+    D_800A9EAC += 4 + (Rng_Rand16() & 7);
+    value = FP_MULTIPLY(shRsin(D_800A9EAC), 10, Q12_SHIFT) - 122;
+    ptr2 = ptr + 461;
+    
+    for (i = 20; i >= 0; i--)
+    {
+        *ptr2-- = value;
+    }
+    
+    for (i = 0; i < 16; i++)
+    {
+        idx = Rng_Rand16() % 21;
+        ptr1[idx] = -1;
+    }
+    
+    for (i = 0; i < 9; i++)
+    {
+        idx = Rng_Rand16() % 21;
+        ptr1[idx] = 0;
+    }
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003BAC4); // 0x8003BAC4
+#endif
+
+void func_8003BBF4() // 0x8003BBF4
+{
+    s32 val;
+    s32 j;
+    s32 i;
+    u8* ptr;
+
+    func_8003BAC4();
+
+    for (i = 0; i < 21; i++)
+    {
+        ptr = &D_800BCDE0[i * 21];
+        
+        for (j = 0; j < 21; j++)
+        {
+            val   = ptr[j + 21];
+            val  += ptr[j - 1];
+            val  += ptr[j];
+            val  += ptr[j + 1];
+            val >>= 2;
+            val  -= 1;
+            
+            if (val <= 0) 
+            {
+                ptr[j] = 0;
+            } 
+            else 
+            {
+                ptr[j] = val;
+            }
+        }
+    }
+}
+
+void func_8003BC8C() // 0x8003BC8C
+{
+    if (D_800A9EB0 == ((D_800A9EB0 / 5) * 5))
+    {
+        func_8003BBF4(D_800A9EB0);
+    }
+
+    D_800A9EB0++;
+    func_8003BA08();
+}
+
+void func_8003BCF4() // 0x8003BCF4
+{
+    s32 i;
+
+    func_8003B7BC();
+
+    for (i = 0; i < 30; i++)
+    {
+        func_8003BBF4();
+    }
+}
+
+// TODO: Remake this whenever we have further context of `D_8002500C`.
+s32 func_8003BD2C() // 0x8003BD2C
+{
+    return ((s32*)D_800BCE18.field_0[0].field_0 - &D_8002500C) >> 2;
+}
+
+void func_8003BD48(s_SubCharacter* chara) // 0x8003BD48
+{
+    u16 var_s0;
+
+    D_800BCE14 = func_80069810();
+    var_s0 = D_800BCE14;
+
+    switch (func_8003BD2C())
+    {
+        case 0:
+            if (chara->position_18.vx >= FP_METER(191.6f) && chara->position_18.vx <= FP_METER(198.8f) && 
+                chara->position_18.vz >= FP_METER(-96.0f) && chara->position_18.vz <= FP_METER(-90.3f))
+            {
+                var_s0 = (var_s0 & ~0x2) | 0x4;
+            }
+            break;
+
+        case 3:
+        case 4:
+            if(chara->position_18.vx >= FP_METER(-100.0f) && chara->position_18.vx <= FP_METER(-94.5f) && 
+               chara->position_18.vz >= FP_METER(-70.3f) && chara->position_18.vz <= FP_METER(-62.0f)) 
+            {
+                var_s0 = (var_s0 & ~0x2) | 0x4;
+            }
+    }
+    
+    func_80069820(var_s0);
+}
+
+void func_8003BE28() // 0x8003BE28
+{
+    func_80069820(D_800BCE14);
+}
+
+s32 func_8003BE50(s32 idx) // 0x8003BE50
+{
+    return &D_800BCE18.field_0[0].field_18[idx]->field_14.field_C;
+}
+
+void GameFs_BgEtcGfxLoad() // 0x8003BE6C
+{
+    Fs_QueueStartReadTim(FILE_TIM_BG_ETC_TIM, FS_BUFFER_1, &D_800A9EB4);
+}
+
+void GameFs_BgItemLoad() // 0x8003BE9C
+{
+    D_800BCE18.field_1BE4.queueIdx_1000 = Fs_QueueStartRead(FILE_BG_BG_ITEM_PLM, &D_800BCE18.field_1BE4);
+}
+
+void func_8003BED0() // 0x8003BED0
+{
+    s_800BE9FC* D_800BE9FC = &D_800BCE18.field_1BE4;
+
+    if (Fs_QueueIsEntryLoaded(D_800BE9FC->queueIdx_1000) == 0 || D_800BE9FC->field_2 != 0)
+    {
+        return;
+    }
+
+    func_800560FC(&D_800BCE18.field_1BE4);
+    func_80056504(&D_800BCE18.field_1BE4, &D_80025528, &D_800A9EBC, 1);
+    func_80056504(&D_800BCE18.field_1BE4, &D_80025530, &D_800A9EC4, 1);
+    func_80056954(&D_800BCE18.field_1BE4);
+}
+
+s32 func_8003BF60(s32 arg0, s32 arg1) 
+{
+    s32 ret;
+    s_800BCE18_0_0_C* ptr;
+    s8 val;
+
+    ret = 0;
+    
+    if (g_SavegamePtr->mapOverlayId_A4 == 0) 
+    {
+        return 1;
+    }
+    
+    if (D_800BCE18.field_0[0].field_0->field_C != NULL) 
+    {
+        ptr = D_800BCE18.field_0[0].field_0->field_C;
+        val = ptr->field_0;
+        
+        if (val != -1)
+        {
+            do 
+            {
+                if (arg0 >= (ptr->field_2 << 8) && (ptr->field_4 << 8) >= arg0 && 
+                    arg1 >= (ptr->field_6 << 8) && (ptr->field_8 << 8) >= arg1 && 
+                    ret < val)
+                {
+                    ret = val;
+                }
+                ptr++;
+                val = ptr->field_0;
+            } while (val != -1);
+        }
+    }
+    return ret;
+}
+
+/** Important for map loading.
+ * Removing it causes the game to get in the loading screen infinitely.
+ */
+void func_8003C048() // 0x8003C048
+{
+    func_80055028();
+    D_800BCE18.field_0[0].field_4 = 0;
+    func_80041C24((s_80041CEC* )0x8016B600, 0x80175600, 0x2C000);
+    func_800697EC();
+    g_SysWork.field_2378 = 0x1000;
+    func_8003EBA0();
+    func_8005B55C(vwGetViewCoord());
+    func_8003CB3C(&D_800BCE18);
+}
+
+void func_8003C0C0() // 0x8003C0C0
+{
+    s_800BCE18_1BAC* ptr = &D_800BCE18.field_1BAC;
+    ptr->field_0         = -1;
+    ptr->field_14        = (s_800BE9FC*)(Fs_GetFileSize(0x58E) + 0x800FE600); // field_14 is defined as a pointer?
+    ptr->field_18        = 0;
+    ptr->field_1C        = 0;
+    ptr->field_20        = 0;
+}
+
+void func_8003C110() // 0x8003C110
+{
+    s32 i;
+    s_800BCE18_0_CC* var_s0;
+    
+    for (i = 0; i < 45; i++) 
+    {
+        if (i != 1) 
+        {
+            D_800BCE18.field_0->field_18[i] = NULL;
+        }
+    } 
+    
+    D_800BCE18.field_0[0].field_14 = Fs_GetFileSize(0x58E) + 0x800FEE00;
+
+    // this part could be rewritten in a less confusing way
+    var_s0 = &D_800BCE18.field_0[0].field_CC;
+    
+    while ((u32)var_s0 < (u32)&D_800BCE18.field_164C) 
+    {
+        func_8003C1AC((u32)var_s0);
+        var_s0 = (int)var_s0 + sizeof(s_800BCE18_0); // matches but fake
+    } 
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003C1AC); // 0x8003C1AC
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003C220); // 0x8003C220
+
+void func_8003C2EC()
+{
+    func_80041FF0();
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003C30C); // 0x8003C30C
+
+void func_8003C368() // 0x8003C368
+{
+    D_800BCE18.field_0[0].field_4 = 1;
+    D_800BCE18.field_0[0].field_8 = g_SysWork.player_4C.chara_0.position_18;
+}
+
+void func_8003C3A0() // 0x8003C3A0
+{
+    D_800BCE18.field_0[0].field_4 = 0;
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003C3AC); // 0x8003C3AC
+
+s32 func_8003C850() // 0x8003C850
+{
+    func_8003C3AC();
+    func_80043740();
+}
+
+void func_8003C878(s32 arg0) // 0x8003C878
+{
+    func_8003CB44(&D_800BCE18);
+
+    while (func_80043830())
+    {
+        func_8003C3AC();
+        Fs_QueueWaitForEmpty();
+    }
+
+    func_80043A24(&g_ObjectTable0[g_ObjectTableIdx], arg0);
+    func_800550D0();
+}
+
+void func_8003C8F8(s_func_8003C8F8* arg0, s8* arg1) // 0x8003C8F8
+{
+    arg0->field_25 = 0;
+    arg0->field_0 = 0;
+
+    func_80056D64((s8*)&arg0->field_16, arg1);
+
+    arg0->field_24 = 0;
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003C92C); // 0x8003C92C
+
+void func_8003CB3C(s_800BCE18* arg0) // 0x8003CB3C
+{
+    arg0->field_2BE8 = 0;
+}
+
+void func_8003CB44(s_800BCE18* arg0) // 0x8003CB44
+{
+    s_800BCE18_2BEC* ptr;
+
+    for (ptr = &arg0->field_2BEC[0]; ptr < &arg0->field_2BEC[arg0->field_2BE8]; ptr++)
+    {
+        func_8003CBA4(ptr);
+    }
+
+    arg0->field_2BE8 = 0;
+}
+
+void func_8003CBA4(s_800BCE18_2BEC* arg0) // 0x8003CBA4
+{
+    GsCOORDINATE2 coord;
+    SVECTOR       vec;
+    MATRIX        mats[2];
+
+    coord.flg   = 0;
+    coord.super = 0;
+
+    coord.coord.t[0] = arg0->gsCoordinate0_4;
+    coord.coord.t[1] = arg0->gsCoordinate1_4;
+    coord.coord.t[2] = arg0->gsCoordinate2_8;
+
+    // Unpack XYZ bitfield (TODO: Was this used anywhere else?)
+    vec.vx = arg0->vx_C << 2;
+    vec.vy = arg0->vy_C;
+    vec.vz = arg0->vz_C << 2;
+
+    func_80096C94(&vec, &coord.coord);
+    func_80049B6C(&coord, &mats[1], &mats[0]);
+    func_8003CC7C(arg0->field_0, &mats[0], &mats[1]);
+}
+
+void func_8003CC7C(s_800BCE18_2BEC_0* arg0, MATRIX* arg1, MATRIX* arg2) // 0x8003CC7C
+{
+    s8                    temp_a0;
+    s_800BCE18_2BEC_0_10* temp_s1;
+    s_800BCE18_2BEC_0_10* temp_s2;
+
+    temp_a0 = arg0->field_10[0].field_9;
+    if (!temp_a0)
+    {
+        return;
+    }
+
+    temp_s2 = arg0->field_8;
+    temp_s1 = arg0->field_10;
+
+    if (temp_a0 >= 3 && temp_a0 < 7)
+    {
+        if (!func_80042C04(temp_a0 - 3))
+        {
+            arg0->field_10[0].field_9 = 0;
+        }
+    }
+
+    if (temp_s1->field_0 != temp_s2->field_0 || temp_s1->field_4 != temp_s2->field_4)
+    {
+        arg0->field_10[0].field_9 = 0;
+        return;
+    }
+
+    func_80057090(arg0, &g_ObjectTable0[g_ObjectTableIdx], 1, arg1, arg2, 0);
+}
+
+s32 func_8003CD5C() // 0x8003CD5C
+{
+    return D_800BCE18.field_1BAC.field_0;
+}
+
+void func_8003CD6C(s_PlayerCombat* arg0) // 0x8003CD6C
+{
+    s32 var_a0;
+    s8  temp_v0;
+
+    temp_v0 = arg0->field_F;
+    var_a0  = NO_VALUE;
+    if (temp_v0 != NO_VALUE)
+    {
+        var_a0 = temp_v0 + 0x80;
+    }
+
+    func_8003CDA0(var_a0);
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003CDA0); // 0x8003CDA0
+
+void func_8003D01C() // 0x8003D01C
+{
+    D_800BCE18.field_1BAC.field_18 &= ~(1 << 31);
+}
+
+void func_8003D03C() // 0x8003D03C
+{
+    D_800BCE18.field_1BAC.field_18 |= 1 << 31;
+}
+
+void func_8003D058() // 0x8003D058
+{
+    MATRIX           mat0;
+    MATRIX           mat1;
+    GsCOORDINATE2*   coord;
+    s_800BCE18_1BAC* ptr0;
+    s_800BE9FC*      ptr1;
+
+    ptr0 = &D_800BCE18.field_1BAC;
+
+    if (ptr0->field_0 != NO_VALUE) 
+    {
+        if (ptr0->field_0 == 164)
+        {
+            coord = &g_SysWork.playerBoneCoords_890[6];
+        } 
+        else 
+        {
+            coord = &g_SysWork.playerBoneCoords_890[10];
+        }
+
+        if (Fs_QueueIsEntryLoaded(ptr0->field_4) != 0) 
+        {
+            ptr1 = ptr0->field_14;
+
+            if (ptr1->field_2 == 0)
+            {
+                func_800560FC(ptr1);
+                func_80056504(ptr1, ptr0->field_8, &ptr0->field_C, 1);
+                func_80056954(ptr1);
+                func_80056C8C(&ptr0->field_18, ptr0->field_14, 0);
+            }
+
+            func_80049B6C(coord, &mat1, &mat0);
+            func_80057090(&ptr0->field_18, &g_ObjectTable0[g_ObjectTableIdx], 1, &mat0, &mat1, 0);
+        }
+    }
+}
+
+void func_8003D160() // 0x8003D160
+{
+    s_FsImageDesc    img;
+    s32              queueIdx;
+    s_800BCE18*      ptr;
+    s_800BCE18_0_CC* ptr2;
+    void*            addr = (void*)0x800FE600;
+
+    func_8003D3BC(&img, 1, 0);
+
+    ptr                               = &D_800BCE18;
+    ptr2                              = &ptr->field_164C;
+    D_800BCE18.field_0[0].field_18[1] = ptr2;
+
+    Fs_QueueStartRead(g_Chara_FileInfo[1].modelFileIdx, addr);
+    queueIdx = Fs_QueueStartReadTim(g_Chara_FileInfo[1].textureFileIdx, FS_BUFFER_1, &img);
+
+    D_800BCE18.field_164C.field_0 = 1;
+    ptr2->field_1                 = 0;
+    ptr2->field_4                 = queueIdx;
+    ptr2->field_8                 = addr;
+    D_800BCE18.field_164C.field_C = img;
+}
+
+s32 func_8003D21C(s_MapOverlayHeader* arg0) // 0x8003D21C
+{
+    s_FsImageDesc    img;
+    s32              j;
+    s32              var_s3;
+    s32              i;
+    s32              ret;
+    s32              ids;
+    s_800BCE18_0_CC* ptr;
+
+    for (ret                           = 0,
+        i                              = 0,
+        D_800BCE18.field_0[0].field_14 = Fs_GetFileSize(0x58E) + 0x800FEE00, 
+        var_s3                         = 0;
+        i < 4; i++)
+    {
+        ids = arg0->charaGroupIds_248[i];
+        ptr = &D_800BCE18.field_0[i].field_CC;
+
+        if (ids != 0) 
+        {
+            if (var_s3 == 0) 
+            {
+                if (ids != ptr->field_0) 
+                {
+                    var_s3 = 1;
+                    for (j = i; j < 4; j++)
+                    {
+                        D_800BCE18.field_0[j].field_CC.field_0 = 0;
+                    }
+                }
+            } 
+
+            if (var_s3 != 0) 
+            {
+                func_8003D3BC(&img, ids, i);
+                ret = func_8003D7D4(ids, i, D_800BCE18.field_0[0].field_14, &img);
+            }
+
+            func_8003D354(&D_800BCE18.field_0[0].field_14, ids);
+        }
+    }
+
+    return ret;
+}
+
+void func_8003D354(s32* arg0, s32 arg1) // 0x8003D354
+{
+    s16 idx;
+    s32 fileSize;
+
+    idx      = g_Chara_FileInfo[arg1].modelFileIdx;
+    fileSize = Fs_GetFileSize(idx);
+    Fs_GetFileSectorAlignedSize(idx);
+    *arg0 += (fileSize + 3) & ~3;
+}
+
+void func_8003D3BC(s_FsImageDesc* img, s32 arg1, s32 arg2) // 0x8003D3BC
+{
+    s16 clutX;
+    s16 clutY;
+    s8  tPage;
+    s8  v;
+    s8  u;
+
+    v = arg1 < 2;
+
+    if (arg1 >= 0 && v)
+    {
+        tPage = 0x1B;
+        v     = 0;
+        u     = 0;
+        clutX = 0x2E0;
+        clutY = 0x1E0;
+    }
+    else 
+    {
+        clutY = 0x1D0;
+        clutX = (arg2 * 0x10) + 0x2C0;
+
+        switch (arg2)
+        {
+            default:
+                arg2 = 0;
+
+            case 0:
+            case 1:
+                tPage = 0x1C;
+                u     = 0;
+                v     = arg2 << 7;
+                break;
+
+            case 2:
+            case 3:
+                tPage = 0x1D;
+                u     = 0;
+                v     = (arg2 - 2) << 7;
+                break;
+        }
+    }
+
+    img->tPage[0] = 0;
+    img->tPage[1] = tPage;
+    img->u        = u;
+    img->v        = v;
+    img->clutX    = clutX;
+    img->clutY    = clutY;
+}
+
+s32 func_8003D444(s32 idx) // 0x8003D444
+{
+    return D_800BCE18.field_0[0].field_18[idx] != 0;
+}
+
+void func_8003D460() {}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003D468); // 0x8003D468
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003D550); // 0x8003D550
+
+// This was matched in decomp.me, but `if (flags & (1 << i))` causes missmatch
+// when inserted.
+#ifdef NON_MATCHING
+void func_8003D5B4(s8 flags) // 0x8003D5B4
+{
+    u8 fileIdx;
+    s32 i;
+    u32 temp;
+    s_800BCE18_0_CC* ptr;
+
+    for (i = 0; i < 4; i++)
+    {
+        ptr = &D_800BCE18.field_0[i].field_CC;
+        if (flags & (1 << i))
+        {
+            func_8003D6A4((u8*)ptr);
+        }
+    }
+
+    i = 0; 
+    
+    D_800BCE18.field_0[0].field_14 = Fs_GetFileSize(FILE_CHARA_HERO_ILM) + 0x800FEE00;
+    
+    for (; i < 4; i++)
+    {
+        ptr = &D_800BCE18.field_0[i].field_CC;
+        
+        fileIdx = ptr->field_0;
+        if (fileIdx != 0)
+        {
+            temp = ptr->field_8 + Fs_GetFileSize(g_Chara_FileInfo[fileIdx].modelFileIdx);
+            if (D_800BCE18.field_0[0].field_14 < temp)
+            {
+                D_800BCE18.field_0[0].field_14 = temp;
+            }
+        }
+    }
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003D5B4); // 0x8003D5B4
+#endif
+
+void func_8003D6A4(u8* idx) // 0x8003D6A4
+{
+    u8 locIdx;
+
+    locIdx = *idx;
+    if (locIdx != 0)
+    {
+        D_800BCE18.field_0[0].field_18[locIdx] = 0;
+        func_8003C1AC((u32)idx);
+    }
+}
+
+void func_8003D6E0(s32 arg0, s32 arg1, void* arg2, s_FsImageDesc* arg3) // 0x8003D6E0
+{
+    s_FsImageDesc img;
+    void*         var_s0;
+
+    if (arg2 != NULL) 
+    {
+        var_s0 = arg2;
+    } 
+    else if (D_800BCE18.field_0[arg1].field_CC.field_0 != 0) 
+    {
+        var_s0 = D_800BCE18.field_0[arg1].field_CC.field_8;
+    } 
+    else 
+    {
+        var_s0 = D_800BCE18.field_0[0].field_14;
+        func_8003D354(&D_800BCE18.field_0[0].field_14, arg0);
+    }
+
+    if (arg3 != NULL) 
+    {
+        img = *arg3;
+    } 
+    else 
+    {
+        func_8003D3BC(&img, arg0, arg1);
+    }
+
+    func_8003D7D4(arg0, arg1, var_s0, &img);
+}
+
+s32 func_8003D7D4(u32 arg0, s32 arg1, void* arg2, s_FsImageDesc* arg3) // 0x8003D7D4
+{
+    s32              queueIdx;
+    s32              idx;
+    s_800BCE18_0_CC* ptr;
+    s_FsImageDesc*   img;
+
+    ptr = &D_800BCE18.field_0[arg1].field_CC;
+    idx = ptr->field_0;
+    img = &ptr->field_C;
+
+    if (arg0 == 0) 
+    {
+        D_800BCE18.field_0[0].field_18[idx] = NULL;
+        return 0;
+    }
+
+    if (idx != 0) 
+    {
+        if (arg0 == idx) 
+        {
+            if (arg2 == ptr->field_8 && memcmp(arg3, img, sizeof(s_FsImageDesc)) == 0)
+            {
+                return 0;
+            }
+        }
+
+        D_800BCE18.field_0[0].field_18[idx] = NULL;
+    }
+
+    D_800BCE18.field_0[0].field_18[arg0] = ptr;
+
+    queueIdx = Fs_QueueStartRead(g_Chara_FileInfo[arg0].modelFileIdx, arg2);
+
+    if (g_Chara_FileInfo[arg0].textureFileIdx != NO_VALUE) 
+    {
+        queueIdx = Fs_QueueStartReadTim(g_Chara_FileInfo[arg0].textureFileIdx, FS_BUFFER_1, arg3);
+    }
+
+    ptr->field_0 = arg0;
+    ptr->field_1 = 0;
+    ptr->field_4 = queueIdx;
+    ptr->field_8 = arg2;
+    ptr->field_C = *arg3;
+
+    return queueIdx;
+}
+
+void func_8003D938() // 0x8003D938
+{
+    func_8003D9C8(&D_800BCE18.field_164C);
+}
+
+void func_8003D95C() // 0x8003D95C
+{
+    s32 temp_a0;
+    s32 i;
+
+    for (i = 0; i < 45; i++)
+    {
+        if (i != 1) 
+        {
+            temp_a0 = D_800BCE18.field_0[0].field_18[i];
+            if (temp_a0 != 0) 
+            {
+                func_8003D9C8(temp_a0);
+            }
+        }
+    }
+}
+
+void func_8003D9C8(s_800BCE18_0_CC* arg0) // 0x8003D9C8
+{
+    s_Skeleton* skel;
+
+    if (arg0->field_1 == 0 && arg0->field_0 != 0 && Fs_QueueIsEntryLoaded(arg0->field_4) != 0)
+    {
+        arg0->field_1 = 1;
+
+        func_800560FC(arg0->field_8);
+        func_80056464(arg0->field_8, g_Chara_FileInfo[arg0->field_0].textureFileIdx, &arg0->field_C, g_Chara_FileInfo[arg0->field_0].field_6_10 & 3);
+
+        skel = &arg0->field_14;
+
+        func_80056954(arg0->field_8);
+        func_80044FE0(skel, &arg0->field_14.field_C, 56); // TODO: Can't fit `s_Bone` at `field_C`. Check `s_Skeleton` size.
+        func_8004506C(skel, arg0->field_8);
+        func_800452EC(skel);
+        func_800453E8(skel, 1);
+    }
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003DA9C); // 0x8003DA9C
+
+s32 func_8003DD74(s32 arg0, s32 arg1) // 0x8003DD74
+{
+    return (arg1 << 10) & 0xFC00;
+}
+
+// TODO: RODATA migration.
+#ifdef NON_MATCHING
+void func_8003DD80(s32 idx, s32 arg1) // 0x8003DD80
+{
+    s_800BCE18_0_CC* temp_a2;
+
+    temp_a2 = D_800BCE18.field_0[0].field_18[idx];
+
+    switch (idx)
+    {
+        case 1:
+            func_8003DE60(&temp_a2->field_14, arg1);
+            break;
+
+        case 7:
+            func_8003E388(&temp_a2->field_14, arg1);
+            break;
+
+        case 26:
+        case 27:
+            func_8003DF84(&temp_a2->field_14, arg1);
+            break;
+
+        case 24:
+            func_8003E08C(&temp_a2->field_14, arg1);
+            break;
+
+        case 30:
+        case 31:
+            func_8003E194(&temp_a2->field_14, arg1);
+            break;
+
+        case 38:
+        case 39:
+            func_8003E238(&temp_a2->field_14, arg1);
+            break;
+
+        case 14:
+            func_8003E414(&temp_a2->field_14, arg1);
+            break;
+
+        case 16:
+            func_8003E4A0(&temp_a2->field_14, arg1);
+            break;
+
+        case 18:
+            func_8003E544(&temp_a2->field_14, arg1);
+            break;
+
+        default:
+            break;
+    }
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003DD80);
+#endif
+
+// TODO: RODATA migration.
+#ifdef NON_MATCHING
+void func_8003DE60(s_Skeleton* skel, s32 arg1) // 0x8003DE60
+{
+    s32 temp_s0;
+
+    temp_s0 = arg1 & 0xF;
+    if (temp_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9ECC, 0);
+
+        switch (temp_s0)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9ED0, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9ED4, 1);
+                break;
+
+            case 3:
+                func_80045468(skel, &D_800A9ED8, 1);
+                break;
+
+            case 4:
+                func_80045468(skel, &D_800A9EDC, 1);
+                break;
+
+            case 5:
+                func_80045468(skel, &D_800A9EE0, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+
+    temp_s0 = arg1 & 0xF0;
+    if (temp_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9EE4, 0);
+
+        switch (temp_s0)
+        {
+            case 16:
+                func_80045468(skel, &D_800A9EE8, 1);
+                break;
+
+            case 32:
+                func_80045468(skel, &D_800A9EEC, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003DE60);
+#endif
+
+void func_8003DF84(s_Skeleton* skel, s32 arg1) // 0x8003DF84
+{
+    s32 temp_v1;
+
+    temp_v1 = arg1 & 0xF;
+    if (temp_v1 != 0)
+    {
+        switch (temp_v1)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9EF4, 0);
+                func_80045468(skel, &D_800A9EF0, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9EF0, 0);
+                func_80045468(skel, &D_800A9EF4, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+
+    temp_v1 = arg1 & 0xF0;
+    if (temp_v1 != 0)
+    {
+        switch (temp_v1)
+        {
+            case 16:
+                func_80045468(skel, &D_800A9EFC, 0);
+                func_80045468(skel, &D_800A9EF8, 1);
+                break;
+
+            case 32:
+                func_80045468(skel, &D_800A9EF8, 0);
+                func_80045468(skel, &D_800A9EFC, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E08C(s_Skeleton* skel, s32 arg1) // 0x8003E08C
+{
+    s32 temp_v1;
+
+    temp_v1 = arg1 & 0xF;
+    if (temp_v1 != 0)
+    {
+        switch (temp_v1)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F04, 0);
+                func_80045468(skel, &D_800A9F00, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F00, 0);
+                func_80045468(skel, &D_800A9F04, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+
+    temp_v1 = arg1 & 0xF0;
+    if (temp_v1 != 0)
+    {
+        switch (temp_v1)
+        {
+            case 16:
+                func_80045468(skel, &D_800A9F0C, 0);
+                func_80045468(skel, &D_800A9F08, 1);
+                break;
+
+            case 32:
+                func_80045468(skel, &D_800A9F08, 0);
+                func_80045468(skel, &D_800A9F0C, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E194(s_Skeleton* skel, s32 arg1) // 0x8003E194
+{
+    s32 temp_s0;
+
+    temp_s0 = arg1 & 0xF;
+    if (temp_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9F10, 0);
+
+        switch (temp_s0)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F14, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F18, 1);
+                break;
+
+            case 3:
+                func_80045468(skel, &D_800A9F1C, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E238(s_Skeleton* skel, s32 arg1) // 0x8003E238
+{
+    s32 var_s0;
+
+    var_s0 = arg1 & 0xF;
+    if (var_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9F20, 0);
+
+        switch (var_s0)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F28, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F2C, 1);
+                break;
+
+            case 3:
+                func_80045468(skel, &D_800A9F30, 1);
+                break;
+
+            case 4:
+                func_80045468(skel, &D_800A9F34, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+
+
+    var_s0 = arg1 & 0xF0;
+    if (var_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9F38, 0);
+
+        switch (var_s0)
+        {
+            case 16:
+                func_80045468(skel, &D_800A9F3C, 1);
+                break;
+
+            case 32:
+                func_80045468(skel, &D_800A9F40, 1);
+                break;
+
+            case 48:
+                func_80045468(skel, &D_800A9F44, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E388(s_Skeleton* skel, s32 arg1) // 0x8003E388
+{
+    s32 temp_a1;
+
+    temp_a1 = arg1 & 0xF;
+    if (temp_a1 != 0)
+    {
+        switch (temp_a1)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F4C, 0);
+                func_80045468(skel, &D_800A9F48, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F48, 0);
+                func_80045468(skel, &D_800A9F4C, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E414(s_Skeleton* skel, s32 arg1) // 0x8003E414
+{
+    s32 temp_a1;
+
+    temp_a1 = arg1 & 0x3;
+    if (temp_a1 != 0)
+    {
+        switch (temp_a1)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F50, 0);
+                func_80045468(skel, &D_800A9F54, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F54, 0);
+                func_80045468(skel, &D_800A9F50, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E4A0(s_Skeleton* skel, s32 arg1) // 0x8003E4A0
+{
+    s32 temp_s0;
+
+    temp_s0 = arg1 & 0xF;
+    if (temp_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9F58, 0);
+
+        switch (temp_s0)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F60, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F64, 1);
+                break;
+
+            case 3:
+                func_80045468(skel, &D_800A9F68, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+void func_8003E544(s_Skeleton* skel, s32 arg1) // 0x8003E544
+{
+    s32 temp_s0;
+
+    temp_s0 = arg1 & 0xF;
+    if (temp_s0 != 0)
+    {
+        func_80045468(skel, &D_800A9F6C, 0);
+
+        switch (temp_s0)
+        {
+            case 1:
+                func_80045468(skel, &D_800A9F74, 1);
+                break;
+
+            case 2:
+                func_80045468(skel, &D_800A9F78, 1);
+                break;
+
+            case 3:
+                func_80045468(skel, &D_800A9F7C, 1);
+                break;
+
+            default:
+                break;
+        }
+    }
+}
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003E5E8);
+
+void GameFs_FlameGfxLoad() // 0x8003E710
+{
+    Fs_QueueStartReadTim(FILE_TIM_FLAME_TIM, FS_BUFFER_1, &D_800A9FA8);
+}
+
+void func_8003E740() // 0x8003E740
+{
+    DVECTOR   sp10;
+    MATRIX    sp18;
+    SVECTOR   sp38;
+    s32       sp40[4];
+    SVECTOR   sp50;
+    DVECTOR   sp58;
+    s32       sp60;
+    s32       temp_a0;
+    s32       temp_s6;
+    s16*      var_a0;
+    s32       i;
+    s32       var_s5;
+    POLY_FT4* poly;
+    s32       temp_s2 = 0;
+
+    if (g_DeltaTime0 != 0)
+    {
+        D_800A9FB0 += 8;
+        for (i = 0; i < 8; i++)
+        {
+            D_800BCDE8[i] = Rng_Rand16();
+        }
+    }
+
+    sp38.vx = 1;
+    sp38.vy = -7;
+    sp38.vz = 0x21;
+    sp38.vx = FP_MULTIPLY(shAngleRegulate(D_800BCDE8[temp_s2++]), 5, Q12_SHIFT) + 1;
+    sp38.vz = FP_MULTIPLY(shAngleRegulate(D_800BCDE8[temp_s2++]), 5, Q12_SHIFT) + 33;
+
+    poly = (POLY_FT4*)GsOUT_PACKET_P;
+
+    func_80049AF8(&g_SysWork.playerBoneCoords_890[PlayerBone_RightHand], &sp18);
+    SetRotMatrix(&sp18);
+    SetTransMatrix(&sp18);
+
+    var_s5 = RotTransPers(&sp38, &sp10, &sp60, &sp60);
+
+    temp_s6  = var_s5 * 4;
+    var_s5 >>= 1;
+    var_s5  -= 2;
+
+    if (var_s5 < 0)
+    {
+        var_s5 = 0;
+    }
+
+    if (temp_s6 >= 0x81 && var_s5 < 0x7FF)
+    {
+        SetPolyFT4(poly);
+        setSemiTrans(poly, 1);
+
+        temp_a0 = D_800BCDE8[temp_s2++];
+
+        if ((temp_a0 & 0xFFF) >= 0xD9A)
+        {
+            D_800A9FB0 -= 0x10 + (temp_a0 & 0xF);
+        }
+
+        if (D_800A9FB0 >= 33)
+        {
+            D_800A9FB0 = 0;
+        }
+
+        setRGB0(poly, D_800A9FB0 + 0x30, D_800A9FB0 + 0x30, D_800A9FB0 + 0x30);
+        poly->tpage = 0x2C;
+        poly->clut  = 0x1032;
+
+        var_a0 = &D_800BCDE8[temp_s2++];
+
+        for (i = 0; i < 4; i++)
+        {
+            sp40[i] = (var_a0[i] & 0xF) - 8;
+        }
+
+        SetRotMatrix(&GsIDMATRIX);
+        SetTransMatrix(&GsIDMATRIX);
+
+        sp50.vz = temp_s6;
+        sp50.vx = sp40[0] - 0x33;
+        sp50.vy = sp40[2] - 0x33;
+
+        RotTransPers(&sp50, &sp58, &sp60, &sp60);
+
+        poly->x0 = sp10.vx + sp58.vx;
+        poly->y0 = sp10.vy + sp58.vy;
+        sp50.vx  = sp40[1] + 0x33;
+        sp50.vy  = sp40[3] - 0x33;
+
+        RotTransPers(&sp50, &sp58, &sp60, &sp60);
+
+        poly->x1 = sp10.vx + sp58.vx;
+        poly->y1 = sp10.vy + sp58.vy;
+        sp50.vx  = -0x33 - sp40[1];
+        sp50.vy  = 0x33 - sp40[3];
+
+        RotTransPers(&sp50, &sp58, &sp60, &sp60);
+
+        poly->x2 = sp10.vx + sp58.vx;
+        poly->y2 = sp10.vy + sp58.vy;
+        sp50.vx  = 0x33 - sp40[0];
+        sp50.vy  = 0x33 - sp40[2];
+
+        RotTransPers(&sp50, &sp58, &sp60, &sp60);
+
+        poly->x3 = sp10.vx + sp58.vx;
+        poly->y3 = sp10.vy + sp58.vy;
+
+        poly->u0 = 0x80;
+        poly->u1 = 0xBF;
+        poly->u2 = 0x80;
+        poly->u3 = 0xBF;
+
+        poly->v0 = 0;
+        poly->v1 = 0;
+        poly->v2 = 0x3F;
+        poly->v3 = 0x3F;
+
+        AddPrim(&g_ObjectTable0[g_ObjectTableIdx].org[var_s5], poly);
+        GsOUT_PACKET_P = (PACKET*)poly + sizeof(POLY_FT4);
+    }
+}
+
+/**
+ * TODO: Please investigate me!
+ */
+void func_8003EB54() // 0x8003EB54
+{
+    g_SysWork.field_2378 = FP_FLOAT_TO(1.0f, Q12_SHIFT);
+
+    g_SysWork.field_235C = &g_SysWork.playerBoneCoords_890[0];
+    g_SysWork.field_236C = &g_SysWork.playerBoneCoords_890[0];
+
+    Math_Vector3Set(&g_SysWork.field_2360, 0, -0x333, -0x2000);
+    Math_SVectorSet(&g_SysWork.field_2370, 0x71, 0, 0);
+}
+
+void func_8003EBA0() // 0x8003EBA0
+{
+    g_SysWork.field_2378 = FP_FLOAT_TO(1.0f, Q12_SHIFT);
+
+    g_SysWork.field_235C = &g_SysWork.playerBoneCoords_890[1];
+    g_SysWork.field_236C = &g_SysWork.playerBoneCoords_890[0];
+
+    Math_Vector3Set(&g_SysWork.field_2360, -0x147, -0x47A, 0x1EB);
+    Math_SVectorSet(&g_SysWork.field_2370, -0xAA, 0, 0);
+}
+
+void func_8003EBF4(s_MapOverlayHeader* arg0) // 0x8003EBF4
+{
+    s_800A9F80* ptr;
+    s32         var_v1;
+    s8          temp_a0;
+    u8          temp_a1;
+
+    temp_a1 = arg0->field_0->field_6;
+
+    var_v1 = 0;
+
+    if (temp_a1 & 4)
+    {
+        var_v1 = (temp_a1 & 3) > 0;
+    }
+
+    switch (arg0->field_16)
+    {
+        case 1:
+            if (var_v1 != 0)
+            {
+                ptr = &D_800A9F84;
+            }
+            else
+            {
+                ptr = &D_800A9F80;
+            }
+            break;
+
+        case 2:
+            if (var_v1 != 0)
+            {
+                ptr = &D_800A9F8C;
+            }
+            else
+            {
+                ptr = &D_800A9F88;
+            }
+            break;
+
+        case 3:
+            ptr = &D_800A9F98;
+            break;
+
+        default:
+            ptr = &D_800A9F80;
+            break;
+    }
+
+    func_8003ED74(ptr->field_0, ptr->field_1);
+}
+
+void func_8003ECBC() // 0x8003ECBC
+{
+    g_SysWork.field_2388.field_15 = 1;
+    g_SavegamePtr->flags_AC      &= ~(1 << 1);
+}
+
+void func_8003ECE4() // 0x8003ECE4
+{
+    g_SysWork.field_2388.field_15 = 0;
+    g_SavegamePtr->flags_AC      |= 1 << 1;
+}
+
+void func_8003ED08() // 0x8003ED08
+{
+    g_SysWork.field_2388.field_15 ^= 1;
+
+    if (g_SysWork.field_2388.field_15 == 1)
+    {
+        g_SavegamePtr->flags_AC &= ~(1 << 1);
+    }
+    else
+    {
+        g_SavegamePtr->flags_AC |= 1 << 1;
+    }
+}
+
+u8 func_8003ED64() // 0x8003ED64
+{
+    return g_SysWork.field_2388.field_15;
+}
+
+void func_8003ED74(s32 arg0, s32 arg1) // 0x8003ED74
+{
+    func_8003EF10(arg0, arg1, 0, 0, 0, 0);
+    func_8003F170();
+}
+
+void func_8003EDA8() // 0x8003EDA8
+{
+    g_SysWork.field_2388.field_14 = 1;
+}
+
+void func_8003EDB8(s32* arg0, s32* arg1) // 0x8003EDB8
+{
+    s_SysWork_2288* ptr0;
+    s_SysWork_2288* ptr1;
+
+    memcpy(arg0, &(ptr0 = &g_SysWork.field_2388)->field_1C[g_SysWork.field_2388.field_15].field_0.field_20.vec_0[1], 4); // Is there a better solution?
+    memcpy(arg1, &(ptr1 = &g_SysWork.field_2388)->field_1C[g_SysWork.field_2388.field_15].field_0.field_24.vec_0[1], 4);
+}
+
+void func_8003EE30(s32 arg0, s8* arg1, s32 arg2, s32 arg3) // 0x8003EE30
+{
+    g_SysWork.field_2388.field_4 = arg1;
+    g_SysWork.field_2388.field_0 = 5;
+    g_SysWork.field_2388.field_8 = arg2;
+    g_SysWork.field_2388.field_C = arg3;
+
+    g_SysWork.field_2388.field_EC[0] = g_SysWork.field_2388.field_1C[0];
+    g_SysWork.field_2388.field_EC[1] = g_SysWork.field_2388.field_1C[1];
+}
+
+void func_8003EEDC(s32 arg0, s32 arg1) // 0x8003EEDC
+{
+    func_8003EF10(arg0, arg1, 0, 0, 0, 0);
+    func_8003F170();
+}
+
+void func_8003EF10(s32 idx0, s32 idx1, s32 arg4, s32 arg5, s32 arg6, s32 arg7) // 0x8003EF10
+{
+    func_8003EF74(&D_800A93CC[idx0], &D_800A93CC[idx1], arg4, arg5, arg6, arg7);
+}
+
+void func_8003EF74(s_sub_StructUnk3* arg0, s_sub_StructUnk3* arg1, s32 arg2, s32 arg3, s32 arg4, s32 arg5) // 0x8003EF74
+{
+    if (arg0 == arg1)
+    {
+        g_SysWork.field_2388.field_16 = 1;
+    }
+    else
+    {
+        g_SysWork.field_2388.field_16 = 0;
+    }
+
+    g_SysWork.field_2388.field_4 = arg3;
+    g_SysWork.field_2388.field_0 = arg2;
+    g_SysWork.field_2388.field_8 = arg4;
+    g_SysWork.field_2388.field_C = arg5;
+
+    g_SysWork.field_2388.field_EC[0] = g_SysWork.field_2388.field_1C[0];
+    g_SysWork.field_2388.field_EC[1] = g_SysWork.field_2388.field_1C[1];
+
+    func_8003F08C(&g_SysWork.field_2388.field_84[0], arg0);
+    func_8003F08C(&g_SysWork.field_2388.field_84[1], arg1);
+}
+
+void func_8003F08C(s_StructUnk3* arg0, s_sub_StructUnk3* arg1) // 0x8003F08C
+{
+    arg0->field_0 = *arg1;
+
+    if (arg1->field_0.s_field_0.field_0 & (1 << 2))
+    {
+        arg0->field_2E = FP_FLOAT_TO(1.0f, Q12_SHIFT);
+    }
+    else
+    {
+        arg0->field_2E = 0;
+    }
+
+    if (arg1->field_0.s_field_0.field_0 & (1 << 4))
+    {
+        arg0->field_2C = FP_FLOAT_TO(1.0f, Q12_SHIFT);
+    }
+    else
+    {
+        arg0->field_2C = 0;
+    }
+
+    switch (arg1->field_E)
+    {
+        case 0:
+        case 1:
+            arg0->field_30 = arg1->field_10;
+            break;
+
+        case 2:
+            arg0->field_30 = 0;
+            break;
+
+        case 3:
+            arg0->field_30 = arg1->field_10;
+            break;
+
+        default:
+            break;
+    }
+}
+
+void func_8003F170() // 0x8003F170
+{
+    MATRIX          sp28;
+    VECTOR          sp48;
+    SVECTOR         sp58;
+    GsCOORDINATE2*  sp60;
+    s32             temp_v0;
+    u8              temp_v1;
+    s32             temp;
+    s_StructUnk3*   ptr2;
+    s_SysWork_2288* ptr = &g_SysWork.field_2388;
+
+    if (g_SysWork.field_2388.field_15 != 0)
+    {
+        g_SysWork.field_2388.field_18 += FP_MULTIPLY((s64)g_DeltaTime0, FP_FLOAT_TO(4.0f, Q12_SHIFT), Q12_SHIFT);
+    }
+    else
+    {
+        g_SysWork.field_2388.field_18 -= FP_MULTIPLY((s64)g_DeltaTime0, FP_FLOAT_TO(4.0f, Q12_SHIFT), Q12_SHIFT);
+    }
+
+    g_SysWork.field_2388.field_18 = CLAMP(g_SysWork.field_2388.field_18, 0, FP_FLOAT_TO(1.0f, Q12_SHIFT));
+
+    if (g_SysWork.field_2388.field_84[g_SysWork.field_2388.field_18 != 0].field_0.field_E == 3)
+    {
+        func_80049AF8(g_SysWork.field_235C, &sp28);
+        ApplyMatrixLV(&sp28, (VECTOR*)&g_SysWork.field_2360, &sp48); // Bug? `g_SysWork.field_2360` is `VECTOR3`.
+        ptr->field_84[g_SysWork.field_2388.field_18 != 0].field_30 = sp48.vz + (sp28.t[2] * 0x10);
+    }
+
+    if (ptr->field_0 == 0)
+    {
+        ptr->field_1C[0] = ptr->field_84[0];
+        ptr->field_1C[1] = ptr->field_84[1];
+    }
+    else
+    {
+        temp_v0 = func_8003F6F0(func_8003F654(ptr), ptr->field_8, ptr->field_C);
+
+        func_8003F838(&ptr->field_1C[0], &ptr->field_EC[0], &ptr->field_84[0], temp_v0);
+        func_8003F838(&ptr->field_1C[1], &ptr->field_EC[1], &ptr->field_84[1], temp_v0);
+
+        if (temp_v0 >= FP_FLOAT_TO(1.0f, Q12_SHIFT))
+        {
+            ptr->field_0 = 0;
+        }
+    }
+
+    func_8003F838(&ptr->field_154, &ptr->field_1C[0], &ptr->field_1C[1], ptr->field_18);
+
+    ptr2 = &ptr->field_154;
+
+    if (ptr->field_14 != 0)
+    {
+        temp_v1       = ptr->field_154.field_0.field_0.s_field_0.field_0;
+        ptr->field_14 = 0;
+
+        if (temp_v1 & 1)
+        {
+            func_8003F08C(ptr2, &D_800A952C);
+        }
+        else if (temp_v1 & 2)
+        {
+            ptr2->field_0.field_4 += 0x4CC;
+        }
+    }
+
+    ptr->field_10 = func_8003FEC0(&ptr2->field_0);
+    func_8003FF2C(ptr2);
+
+    temp = FP_MULTIPLY(func_8003F4DC(&sp60, &sp58, ptr2->field_0.field_4, ptr2->field_0.field_0.s_field_0.field_2, func_80080A10(), &g_SysWork), g_SysWork.field_2378, Q12_SHIFT);
+
+    func_800554C4(temp, ptr2->field_2C, sp60, g_SysWork.field_235C, &sp58, g_SysWork.field_2360.vx, g_SysWork.field_2360.vy, g_SysWork.field_2360.vz, D_800BCE18.field_0[0].field_0->field_8);
+    func_80055814(ptr2->field_30);
+
+    if (ptr->field_154.field_0.field_0.s_field_0.field_0 & (1 << 3))
+    {
+        func_8003E740();
+    }
+}
+
+// TODO: RODATA migration.
+#ifdef NON_MATCHING
+s32 func_8003F4DC(GsCOORDINATE2** arg0, SVECTOR* rot, s32 arg2, s32 arg3, u32 arg4, s_SysWork* sysWork) // 0x8003F4DC
+{
+    s32     temp;
+    s32     res;
+    SVECTOR vec;
+
+    if (arg3 != 2)
+    {
+        arg4 = 1;
+    }
+
+    res = arg2;
+    if (arg4 == 0)
+    {
+        res = 0;
+    }
+
+    switch (arg4)
+    {
+        default:
+        case 1:
+            *arg0 = sysWork->field_236C;
+            break;
+
+        case 0:
+        case 2:
+        case 3:
+        case 4:
+        case 5:
+            *arg0 = NULL;
+            break;
+    }
+
+    switch (arg4)
+    {
+        default:
+        case 1:
+            vec = sysWork->field_2370;
+            break;
+
+        case 0:
+            vec.vx = 0;
+            vec.vy = 0xFC00;
+            vec.vz = 0;
+            break;
+
+        case 2:
+            vec.vx = 0xFF1D;
+            vec.vy = 0x8AA;
+            vec.vz = 0;
+            break;
+
+        case 3:
+            vec.vx = 0xFF1D;
+            vec.vy = 0xFCAB;
+            vec.vz = 0;
+            break;
+
+        case 4:
+            vec.vx = 0xFF1D;
+            vec.vy = 0xAA;
+            vec.vz = 0;
+            break;
+
+        case 5:
+            vec.vx = 0xFF1D;
+            vec.vy = 0x4AA;
+            vec.vz = 0;
+            break;
+    }
+
+    rot->vy = -shRsin(vec.vx);
+    temp    = shRcos(vec.vx);
+    rot->vz = FP_MULTIPLY(temp, shRcos(vec.vy), Q12_SHIFT);
+    rot->vx = FP_MULTIPLY(temp, shRsin(vec.vy), Q12_SHIFT);
+    return res;
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003F4DC);
+#endif
+
+// TODO: RODATA migration.
+#ifdef NON_MATCHING
+u32 func_8003F654(s_func_8003F654* arg0)
+{
+    switch (arg0->field_0)
+    {
+        default:
+        case PrimitiveType_None:
+            break;
+
+        case PrimitiveType_S8:
+            return *arg0->field_4;
+
+        case PrimitiveType_U8:
+            return *(u8*)arg0->field_4;
+
+        case PrimitiveType_S16:
+            return *(s16*)arg0->field_4;
+
+        case PrimitiveType_U16:
+            return *(u16*)arg0->field_4;
+
+        case PrimitiveType_S32:
+            return *(s32*)arg0->field_4;
+    }
+
+    return 0;
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003F654); // 0x8003F654
+#endif
+
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003F6F0); // 0x8003F6F0
+
+s32 Math_GetWeightedAverage(s32 a, s32 b, s32 weight) // 0x8003F7E4
+{
+    return Math_MulFixed(a, FP_ALPHA(1.0f) - weight, Q12_SHIFT) + Math_MulFixed(b, weight, Q12_SHIFT);
+}
+
+void func_8003F838(s_StructUnk3* arg0, s_StructUnk3* arg1, s_StructUnk3* arg2, s32 weight) // 0x8003F838
+{
+    s32 weight0;
+    s32 weight1;
+    s32 weight2;
+    u32 temp;
+
+    weight0 = weight * 2;
+    weight0 = CLAMP(weight0, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+    weight1 = (weight - FP_ALPHA(0.5f)) * 2;
+    weight1 = CLAMP(weight1, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+
+    if (weight < FP_ALPHA(0.5f))
+    {
+        arg0->field_0.field_0.s_field_0.field_0 = arg1->field_0.field_0.s_field_0.field_0;
+    }
+    else
+    {
+        arg0->field_0.field_0.s_field_0.field_0 = arg2->field_0.field_0.s_field_0.field_0;
+    }
+
+    func_8003FCB0(&arg0->field_0, &arg1->field_0, &arg2->field_0, weight);
+
+    if (arg1->field_2C == FP_ALPHA(0.0f))
+    {
+        arg0->field_2C = Math_GetWeightedAverage(0, arg2->field_2C, weight1);
+    }
+    else
+    {
+        arg0->field_2C = Math_GetWeightedAverage(arg1->field_2C, arg2->field_2C, weight0);
+    }
+
+    if (arg1->field_0.field_0.s_field_0.field_0 & (1 << 0))
+    {
+        if (arg2->field_0.field_0.s_field_0.field_0 & (1 << 0))
+        {
+            arg0->field_0.field_0.s_field_0.field_1 = Math_GetWeightedAverage(arg1->field_0.field_0.s_field_0.field_1, arg2->field_0.field_0.s_field_0.field_1, weight);
+        }
+        else
+        {
+            arg0->field_0.field_0.s_field_0.field_1 = Math_GetWeightedAverage(arg1->field_0.field_0.s_field_0.field_1, arg2->field_0.field_0.s_field_0.field_1, weight1);
+        }
+    }
+    else
+    {
+        if (arg2->field_0.field_0.s_field_0.field_0 & (1 << 0))
+        {
+            arg0->field_0.field_0.s_field_0.field_1 = Math_GetWeightedAverage(arg1->field_0.field_0.s_field_0.field_1, arg2->field_0.field_0.s_field_0.field_1, weight0);
+        }
+        else
+        {
+            arg0->field_0.field_0.s_field_0.field_1 = Math_GetWeightedAverage(arg1->field_0.field_0.s_field_0.field_1, arg2->field_0.field_0.s_field_0.field_1, weight);
+        }
+    }
+
+    if (arg1->field_0.field_E == 0)
+    {
+        if (arg2->field_0.field_E != 0)
+        {
+            arg0->field_0.field_E = arg2->field_0.field_E;
+            func_8003FD38(arg0, arg1, arg2, weight, weight0, weight1);
+        }
+        else
+        {
+            temp                  = arg2->field_0.field_E;
+            arg0->field_0.field_E = temp;
+            func_8003FD38(arg0, arg1, arg2, weight, weight, weight);
+        }
+    }
+    else if (arg2->field_0.field_E == 0)
+    {
+        if (weight1 >= FP_ALPHA(1.0f))
+        {
+            arg0->field_0.field_E = arg2->field_0.field_E;
+        }
+        else
+        {
+            arg0->field_0.field_E = arg1->field_0.field_E;
+        }
+        func_8003FD38(arg0, arg1, arg2, weight, weight1, weight0);
+    }
+    else
+    {
+        temp                  = arg2->field_0.field_E;
+        arg0->field_0.field_E = temp;
+        func_8003FD38(arg0, arg1, arg2, weight, weight, weight);
+    }
+
+    arg0->field_0.field_8 = Math_GetWeightedAverage(arg1->field_0.field_8, arg2->field_0.field_8, weight);
+    arg0->field_0.field_A = Math_GetWeightedAverage(arg1->field_0.field_A, arg2->field_0.field_A, weight);
+    arg0->field_0.field_C = Math_GetWeightedAverage(arg1->field_0.field_C, arg2->field_0.field_C, weight);
+
+    if (arg1->field_0.field_0.s_field_0.field_2 == 1 && arg2->field_0.field_0.s_field_0.field_2 == 2)
+    {
+        if (weight < FP_ALPHA(5.0f / 6.0f))
+        {
+            weight2                                 = FP_MULTIPLY(weight, FP_ALPHA(1.2f), Q12_SHIFT);
+            weight2                                 = CLAMP(weight2, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+            arg0->field_0.field_0.s_field_0.field_2 = arg1->field_0.field_0.s_field_0.field_2;
+            arg0->field_0.field_4                   = Math_GetWeightedAverage(arg1->field_0.field_4, 0, weight2);
+        }
+        else
+        {
+            weight2                                 = (weight - FP_ALPHA(5.0f / 6.0f)) * 6;
+            weight2                                 = CLAMP(weight2, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+            arg0->field_0.field_0.s_field_0.field_2 = arg2->field_0.field_0.s_field_0.field_2;
+            weight0                                 = arg2->field_0.field_4;
+            arg0->field_0.field_4                   = Math_GetWeightedAverage(FP_ALPHA(0.0f), weight0, weight2);
+        }
+    }
+    else if (arg1->field_0.field_0.s_field_0.field_2 == 2 && arg2->field_0.field_0.s_field_0.field_2 == 1)
+    {
+        if (weight < FP_ALPHA(1.0f / 6.0f))
+        {
+            weight2                                 = weight * 6;
+            weight2                                 = CLAMP(weight2, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+            arg0->field_0.field_0.s_field_0.field_2 = arg1->field_0.field_0.s_field_0.field_2;
+            arg0->field_0.field_4                   = Math_GetWeightedAverage(arg1->field_0.field_4, FP_ALPHA(0.0f), weight2);
+        }
+        else
+        {
+            weight2                                 = FP_MULTIPLY(weight - FP_ALPHA(1.0f / 6.0f), FP_ALPHA(1.2f), Q12_SHIFT);
+            weight2                                 = CLAMP(weight2, FP_ALPHA(0.0f), FP_ALPHA(1.0f));
+            arg0->field_0.field_0.s_field_0.field_2 = arg2->field_0.field_0.s_field_0.field_2;
+            arg0->field_0.field_4                   = Math_GetWeightedAverage(FP_ALPHA(0.0f), arg2->field_0.field_4, weight2);
+        }
+    }
+    else
+    {
+        if (arg1->field_0.field_0.s_field_0.field_2 != 0 && arg2->field_0.field_0.s_field_0.field_2 == 0)
+        {
+            if (weight >= FP_ALPHA(1.0f))
+            {
+                arg0->field_0.field_0.s_field_0.field_2 = arg2->field_0.field_0.s_field_0.field_2;
+            }
+            else
+            {
+                arg0->field_0.field_0.s_field_0.field_2 = arg1->field_0.field_0.s_field_0.field_2;
+            }
+        }
+        else
+        {
+            arg0->field_0.field_0.s_field_0.field_2 = arg2->field_0.field_0.s_field_0.field_2;
+        }
+
+        arg0->field_0.field_4 = Math_GetWeightedAverage(arg1->field_0.field_4, arg2->field_0.field_4, weight);
+    }
+
+    if (arg1->field_0.field_18.vec_0[0] == 0 && arg2->field_0.field_18.vec_0[0] != 0)
+    {
+        func_8003FE04(&arg0->field_0, &arg1->field_0, &arg2->field_0, weight1);
+    }
+    else
+    {
+        func_8003FE04(&arg0->field_0, &arg1->field_0, &arg2->field_0, weight);
+    }
+}
+
+void func_8003FCB0(s_sub_StructUnk3* arg0, s_sub_StructUnk3* arg1, s_sub_StructUnk3* arg2, s32 arg3) // 0x8003FCB0
+{
+    s32 p0;
+    
+    p0 = FP_ALPHA(1.0f) - arg3;
+    LoadAverageCol(&arg1->field_20.vec_0[1], &arg2->field_20.vec_0[1], p0, arg3, &arg0->field_20.vec_0[1]);
+    LoadAverageCol(&arg1->field_24.vec_0[1], &arg2->field_24.vec_0[1], p0, arg3, &arg0->field_24.vec_0[1]);
+}
+
+void func_8003FD38(s_StructUnk3* arg0, s_StructUnk3* arg1, s_StructUnk3* arg2, s32 weight0, s32 weight1, s32 alphaTo) // 0x8003FD38
+{
+    if (arg1->field_2E != arg2->field_2E)
+    {
+        arg0->field_2E = Math_GetWeightedAverage(arg1->field_2E, arg2->field_2E, weight0);
+    }
+    else
+    {
+        arg0->field_2E = arg2->field_2E;
+    }
+
+    arg0->field_30         = Math_GetWeightedAverage(arg1->field_30, arg2->field_30, weight0);
+    arg0->field_0.field_10 = Math_GetWeightedAverage(arg1->field_0.field_10, arg2->field_0.field_10, weight1);
+    arg0->field_0.field_6  = Math_GetWeightedAverage(arg1->field_0.field_6, arg2->field_0.field_6, weight0);
+
+    LoadAverageCol(arg1->field_0.field_14.vec_0, arg2->field_0.field_14.vec_0, FP_ALPHA(1.0f) - alphaTo, alphaTo, arg0->field_0.field_14.vec_0);
+}
+
+void func_8003FE04(s_sub_StructUnk3* arg0, s_sub_StructUnk3* arg1, s_sub_StructUnk3* arg2, s32 alphaTo) // 0x8003FE04
+{
+    s32 alphaFrom;
+
+    alphaFrom = FP_ALPHA(1.0f) - alphaTo;
+    LoadAverageCol(&arg1->field_18.vec_0[1], &arg2->field_18.vec_0[1], alphaFrom, alphaTo, &arg0->field_18.vec_0[1]);
+    LoadAverageCol(&arg1->field_1C.vec_0[1], &arg2->field_1C.vec_0[1], alphaFrom, alphaTo, &arg0->field_1C.vec_0[1]);
+
+    if ((arg0->field_18.field_0 & ~0xFF) || (arg0->field_1C.field_0 & ~0xFF))
+    {
+        arg0->field_18.vec_0[0] = 1;
+    }
+    else
+    {
+        arg0->field_18.vec_0[0] = 0;
+    }
+}
+
+// TODO: Requires `D_800C4168` to be `const`, but `func_800553C4` writes to it so it can't be?
+#ifdef NON_MATCHING
+s32 func_8003FEC0(s_sub_StructUnk3* arg0) // 0x8003FEC0
+{
+    if (D_800C4168.field_1 != 0)
+    {
+        return arg0->field_10;
+    }
+
+    if (D_800C4168.field_0 == 1)
+    {
+        return vwOresenHokan(D_800A9FB4, 5, arg0->field_4, 0, FP_FLOAT_TO(2.0f, Q12_SHIFT));
+    }
+
+    return FP_FLOAT_TO(20.0f, Q12_SHIFT);
+}
+#else
+INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8003AB28", func_8003FEC0);
+#endif
+
+void func_8003FF2C(s_StructUnk3* arg0) // 0x8003FF2C
+{
+    s32 temp_a0;
+    s32 temp_v1;
+    s32 var_t0;
+
+    temp_v1 = FP_MULTIPLY(arg0->field_2E, ((g_GameWork.config_0.optBrightness_22 * 8) + 4), Q12_SHIFT);
+    var_t0  = CLAMP(temp_v1, 0, 0xFF);
+
+    func_80055330(arg0->field_0.field_0.s_field_0.field_2, arg0->field_0.field_6, arg0->field_0.field_0.s_field_0.field_1, arg0->field_0.field_8, arg0->field_0.field_A, arg0->field_0.field_C, var_t0);
+    func_800553C4(arg0->field_0.field_E != 0, arg0->field_0.field_14.vec_0[0], arg0->field_0.field_14.vec_0[1], arg0->field_0.field_14.vec_0[2]);
+    temp_a0 = arg0->field_0.field_10;
+    func_80055840(temp_a0, temp_a0 + FP_FLOAT_TO(1.0f, Q12_SHIFT));
+    func_800553E0(arg0->field_0.field_18.vec_0[0], arg0->field_0.field_18.vec_0[1], arg0->field_0.field_18.vec_0[2], arg0->field_0.field_18.vec_0[3], arg0->field_0.field_1C.vec_0[1], arg0->field_0.field_1C.vec_0[2], arg0->field_0.field_1C.vec_0[3]);
+}
+
+void func_80040004(s_800BCE18* arg0) // 0x80040004
+{
+    D_800BCE18.field_1BD8 = &arg0->field_0[2].field_26C;
+}
+
+void func_80040014() // 0x80040014
+{
+    func_80069860(g_SysWork.player_4C.chara_0.position_18.vx, g_SysWork.player_4C.chara_0.position_18.vz, D_800BCE18.field_1BD8);
+};
